# ResponseCaching: started conversion to pipes