--- a/src/Middleware/perf/ResponseCaching.Microbenchmarks/AssemblyInfo.cs
+++ b/src/Middleware/perf/ResponseCaching.Microbenchmarks/AssemblyInfo.cs
@@ -1,0 +1,1 @@
+[assembly: BenchmarkDotNet.Attributes.AspNetCoreBenchmark]