… src/Middleware/ResponseCaching/src/Streams/SegmentReadStream.cs
… src/Middleware/ResponseCaching/test/ResponseCachingMiddlewareTests.cs
… src/Middleware/ResponseCaching/test/SegmentReadStreamTests.cs