I've started to look into migrating the `ResponseCaching` middleware to pipes.

For now this only use pipes to write the response from the cache.
This already allow to remove some code.
It also remove a memory copy when saving to the cache, however I do not understand for which scenario it was needed before (and so it would be still needed now) or if it was just a consequence of the design.

Migrating the capturing of the response to cache, getting read of the stream implementation, is more tricky and I am not sure what it would be a correct design.
Currently it rely on replacing the `Body` with a shim of the response stream that capture all the writes and duplicate them to buffer, however it is not possible to replace the `BodyWriter` instead.
Also, should not the shim be applied on the `IHttpResponseBodyFeature`, like it is done in the `ResponseCompression` middleware?
I imagine would make sense to have a common design for both middlewares.

I will keep looking at this, but I opened this draft PR so if you already have a design in mind we can discuss it.
I don't think merging only this part makes sense.

Related #11377