@alefranz 2019-11-12
I had another look at this and it looks like the pipe writer is built around the stream, so keep wrapping the original stream doesn't give a performance disadvantage.
What am I missing?