@anurse 2019-11-18
@jkotalik can you take a look at this and see what you think of the design so we know how to move forward?