@alefranz 2019-12-02
> It definitely is possible to replace the BodyWriter instead, however I can see that there would be some difficulties.

Hi @jkotalik ,
it would also require to make to change the `HttpResponse` to be able to replace the `BodyWriter` which currently is a readonly property.
Do you know if there is any benchmark I can use as a base to find out if there could be a significant performance benefit in not using the stream?

About this PR, I've rebased it and addressed the feedback.
I will mark it for review once I'm done with the tests.