@alefranz 2019-12-03
I've also added a benchmark and it seems like slightly slower then before although I'm not sure the benchmark is really representative and the error margin is quite high, 

Before:

|         Method |      Mean |     Error |    StdDev |    Median |      Op/s |  Gen 0 | Allocated |
|--------------- |----------:|----------:|----------:|----------:|----------:|-------:|----------:|
|          Cache | 11.307 us | 0.1968 us | 0.4562 us | 11.492 us |  88,443.0 |      - |   4.73 KB |
| ServeFromCache |  2.884 us | 0.0157 us | 0.0139 us |  2.882 us | 346,758.8 | 0.0267 |   2.37 KB |

After:

|         Method |      Mean |     Error |    StdDev |      Op/s |  Gen 0 | Allocated |
|--------------- |----------:|----------:|----------:|----------:|-------:|----------:|
|          Cache | 12.086 us | 1.5640 us | 4.5375 us |  82,737.1 |      - |   4.16 KB |
| ServeFromCache |  3.206 us | 0.0259 us | 0.0216 us | 311,938.1 | 0.0267 |   2.45 KB |