@anurse 2019-12-04
> I've also added a benchmark and it seems like slightly slower then before

Hmmmmmmmm... that's not ideal. Given that we're already a little unsure of the value of this change I'm not sure it's worth it right now.