@alefranz 2019-12-08
I've reworked the benchmark, using an actual `Pipe` which should be more representative of the behavior with Kestrel.
I've also added different response sizes.
Now it results in a speedup and less allocation except for really small payloads of a few bytes.
I've also added a large payload (1MB) even if the default limit is 64KB.

It would be great if someone can review the benchmark to confirm the validity of these results:
https://github.com/aspnet/AspNetCore/pull/16961/files#diff-3ec34e95e14d0484ca182d967bb33d4c

Before:

|         Method |    Size |         Mean |      Error |     StdDev |      Op/s |   Gen 0 |   Gen 1 |  Gen 2 | Allocated |
|--------------- |-------- |-------------:|-----------:|-----------:|----------:|--------:|--------:|-------:|----------:|
|          Cache |     100 |    15.897 us |  0.3242 us |  0.9353 us |  62,903.5 |  0.0610 |       - |      - |   6.06 KB |
| ServeFromCache |     100 |     6.839 us |  0.1818 us |  0.5333 us | 146,213.5 |  0.0381 |       - |      - |   3.08 KB |
|          Cache |   65536 |    85.582 us |  1.6837 us |  1.2174 us |  11,684.6 |  4.2725 |  0.6104 |      - | 314.45 KB |
| ServeFromCache |   65536 |   107.485 us |  2.0812 us |  2.3133 us |   9,303.6 |       - |       - |      - |   5.12 KB |
|          Cache | 1048576 | 1,190.080 us | 23.0963 us | 31.6145 us |     840.3 | 46.8750 | 25.3906 | 3.9063 | 841.38 KB |
| ServeFromCache | 1048576 |   296.630 us |  5.8470 us |  5.7425 us |   3,371.2 |       - |       - |      - |   5.46 KB |

After:

|         Method |    Size |       Mean |      Error |     StdDev |      Op/s |   Gen 0 |   Gen 1 |  Gen 2 | Allocated |
|--------------- |-------- |-----------:|-----------:|-----------:|----------:|--------:|--------:|-------:|----------:|
|          Cache |     100 |  16.590 us |  0.3300 us |  0.8280 us |  60,279.1 |  0.0610 |       - |      - |    5.4 KB |
| ServeFromCache |     100 |   7.091 us |  0.1401 us |  0.3133 us | 141,015.8 |  0.0305 |       - |      - |   2.97 KB |
|          Cache |   65536 |  72.965 us |  1.4504 us |  2.7242 us |  13,705.2 |  2.6855 |  0.4883 |      - | 188.16 KB |
| ServeFromCache |   65536 |  88.289 us |  1.9925 us |  2.5199 us |  11,326.5 |       - |       - |      - |   3.61 KB |
|          Cache | 1048576 | 934.144 us | 18.6188 us | 19.1201 us |   1,070.5 | 31.2500 | 17.5781 | 3.9063 | 832.39 KB |
| ServeFromCache | 1048576 | 249.305 us |  4.4989 us |  3.9881 us |   4,011.1 |       - |       - |      - |   3.61 KB |