@anurse 2020-01-06
Next steps: @JunTaoLuo 's going to clone and take a look and do some checking here, as well as checking the benchmark validity.

We'll also need to rebase ;). @alefranz can you do that?

The benchmark results look OK to me, so it seems like it's worth taking if only so that we use more pipes.