@alefranz 2020-01-12
Addressed feedback and rebased.
It should be good to merge now.

Thank you!