@JunTaoLuo 2020-01-13
Looks like a test failure:
```
A total of 1 test files matched the specified pattern.
[xUnit.net 00:00:01.41]     Microsoft.AspNetCore.ResponseCaching.Tests.CachedResponseBodyTests.Copy_DoNothingWhenNoSegments [FAIL]
  X Microsoft.AspNetCore.ResponseCaching.Tests.CachedResponseBodyTests.Copy_DoNothingWhenNoSegments [523ms]
  Error Message:
   System.OperationCanceledException : The operation was canceled.
  Stack Trace:
     at System.Threading.CancellationToken.ThrowOperationCanceledException()
   at System.Threading.CancellationToken.ThrowIfCancellationRequested()
   at System.IO.Pipelines.Pipe.GetReadAsyncResult()
   at System.IO.Pipelines.Pipe.DefaultPipeReader.GetResult(Int16 token)
   at Microsoft.AspNetCore.ResponseCaching.Tests.CachedResponseBodyTests.ReceiveDataAsync(PipeReader reader, List`1 receivedSegments, CancellationToken cancellationToken)
   at Microsoft.AspNetCore.ResponseCaching.Tests.CachedResponseBodyTests.Copy_DoNothingWhenNoSegments() in /_/src/Middleware/ResponseCaching/test/CachedResponseBodyTests.cs:line 52
--- End of stack trace from previous location ---
Results File: /datadisks/disk1/work/A9BF0946/w/B1FF0983/e/TestResults/_c001LHN_2020-01-12_13_45_39.trx
```