@alefranz 2020-01-17
Hi John, apologies for the delay.
I'm confident that the issue was caused by a timeout too short, which I have now increased.